% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesroc-package.R
\docType{package}
\name{bayesroc-package}
\alias{bayesroc}
\alias{bayesroc-package}
\title{bayesroc: Bayesian ROC-Based Signal Detection Theory Models}
\description{
Generates Stan or NumPyro code from R formula syntax for Bayesian ordinal signal detection models. Fits models via 'cmdstanr' or JAX/NumPyro backends.
}
\author{
\strong{Maintainer}: Jedidiah Whitridge \email{jww828@mun.ca}

}
\keyword{internal}
