% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{c.broc_prior}
\alias{c.broc_prior}
\title{Combine Multiple Priors}
\usage{
\method{c}{broc_prior}(...)
}
\arguments{
\item{...}{Prior specifications from broc_prior()}
}
\value{
A list of prior specifications
}
\description{
Combine Multiple Priors
}
