% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broc_model.R
\name{get_stan_code}
\alias{get_stan_code}
\title{Extract Generated Stan Code}
\usage{
get_stan_code(x)
}
\arguments{
\item{x}{A \code{broc_model} object created by \code{\link[=broc]{broc()}}.}
}
\value{
A character string of class \code{stan_code} containing the generated Stan program.
}
\description{
Returns the Stan program generated by \code{\link[=broc]{broc()}} for inspection or manual use.
}
\examples{
\dontrun{
model <- broc(brf(conf | old ~ cond), data = dat, family = evsd())
get_stan_code(model)
}
}
\seealso{
\code{\link[=get_stan_data]{get_stan_data()}} for the matching data list; \code{\link[=get_numpyro_config]{get_numpyro_config()}}
for the equivalent artifact used by the JAX backend.
}
