% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broc_model.R
\name{get_stan_data}
\alias{get_stan_data}
\title{Extract Stan Data List}
\usage{
get_stan_data(x)
}
\arguments{
\item{x}{A \code{broc_model} object created by \code{\link[=broc]{broc()}}.}
}
\value{
A named list of data values formatted for Stan.
}
\description{
Returns the data list that would be passed to Stan's \verb{$sample()} method.
}
\examples{
\dontrun{
model <- broc(brf(conf | old ~ cond), data = dat, family = evsd())
str(get_stan_data(model))
}
}
\seealso{
\code{\link[=get_stan_code]{get_stan_code()}} for the generated program; \code{\link[=get_numpyro_config]{get_numpyro_config()}}
for the equivalent artifact used by the JAX backend.
}
