% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{link_probit}
\alias{link_probit}
\title{Probit Link Function (default)}
\usage{
link_probit()
}
\value{
A \code{broc_link} object to pass to a family's \code{link} argument.
}
\description{
Uses the standard normal CDF.
}
