% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_smooths.R
\name{plot.conditional_smooths}
\alias{plot.conditional_smooths}
\title{Plot Conditional Smooths}
\usage{
\method{plot}{conditional_smooths}(x, ...)
}
\arguments{
\item{x}{A \code{conditional_smooths} object}

\item{...}{Additional arguments passed to ggplot}
}
\value{
A ggplot object
}
\description{
Plot Conditional Smooths
}
\keyword{internal}
