% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{print.broc_link}
\alias{print.broc_link}
\title{Print method for a bayesroc link}
\usage{
\method{print}{broc_link}(x, ...)
}
\arguments{
\item{x}{The object to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for a bayesroc link
}
