% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broc_model.R
\name{print.broc_model}
\alias{print.broc_model}
\title{Print Method for SDT Model}
\usage{
\method{print}{broc_model}(x, ...)
}
\arguments{
\item{x}{The object to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for SDT Model
}
