% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.broc_prior_list}
\alias{print.broc_prior_list}
\title{Print Method for Prior List}
\usage{
\method{print}{broc_prior_list}(x, ...)
}
\arguments{
\item{x}{The object to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for Prior List
}
