% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.broc_prior_table}
\alias{print.broc_prior_table}
\title{Print method for get_broc_prior output}
\usage{
\method{print}{broc_prior_table}(x, ...)
}
\arguments{
\item{x}{The object to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for get_broc_prior output
}
