% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{print.prior_lookup}
\alias{print.prior_lookup}
\title{Print Method for Prior Lookup}
\usage{
\method{print}{prior_lookup}(x, ...)
}
\arguments{
\item{x}{The object to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for Prior Lookup
}
