% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pp_check}
\alias{loo}
\alias{loo_compare}
\alias{posterior_epred}
\alias{posterior_linpred}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayesplot}{\code{\link[bayesplot]{pp_check}}}

  \item{loo}{\code{\link[loo]{loo}}, \code{\link[loo]{loo_compare}}}

  \item{rstantools}{\code{\link[rstantools]{posterior_epred}}, \code{\link[rstantools]{posterior_linpred}}}
}}

